fix: 폴더 구조 수정 및 SRP
