+++ Yut/Yut/Features/ARGamePlay/PlayView.swift
@@ -2,7 +2,7 @@ import SwiftUI
 import RealityKit
 import ARKit
 
-struct ContentView : View {
+struct PlayView : View {
     // 상태 관리 객체 (AR의 현재 단계, 명령 스트림, 윷 결과 등 공유)
     @StateObject var arState = ARState()
     
@@ -14,7 +14,7 @@ struct ContentView : View {
             
             VStack {
                 // 게임 상태에 따라 상단 안내 뷰 + 하단 인터랙션 UI를 함께 표시
-                switch arState.currentState {
+                switch arState.gamePhase {
                     
                 // 1. 바닥 탐색 중 (아직 충분히 인식되지 않음)
                 case .searchingForSurface:
@@ -24,29 +24,20 @@ struct ContentView : View {
                         minRequiredArea: arState.minRequiredArea
                     )
                     Spacer()
-                    // 다음 버튼 비활성화
-                    RoundedBrownButton(title: "다음", isEnabled: false) { }
-
-                // 2. 평면 인식 완료됨 (다음 단계로 넘어갈 수 있음)
-                case .completedSearching:
-                    ProgressBar(
-                        text: "바닥을 충분히 색칠해 주세요.",
-                        currentProgress: arState.recognizedArea,
-                        minRequiredArea: arState.minRequiredArea
-                    )
-                    Spacer()
-                    // 다음 버튼 활성화 → 보드 배치 단계로 전환
-                    RoundedBrownButton(title: "다음", isEnabled: true) {
-                        arState.currentState = .placeBoard
+                    let canProceed = arState.recognizedArea >= arState.minRequiredArea
+                    RoundedBrownButton(title: "다음", isEnabled: canProceed) {
+                        if canProceed {
+                            arState.gamePhase = .placeBoard
+                        }
                     }
-
-                // 3. 사용자가 탭으로 보드를 놓는 단계
+                    
+                // 2. 사용자가 탭으로 보드를 놓는 단계
                 case .placeBoard:
                     InstructionView(text: "탭해서 말판을 배치하세요.")
                     Spacer()
                     EmptyView() // 버튼 없음
 
-                // 4. 핀치/드래그로 위치/크기 조정 단계
+                // 3. 핀치/드래그로 위치/크기 조정 단계
                 case .adjustingBoard:
                     InstructionView(text: "핀치와 드래그로\n보드의 크기와 위치를 조정하세요.")
                     Spacer()
@@ -54,18 +45,18 @@ struct ContentView : View {
                     RoundedBrownButton(title: "여기에 배치", isEnabled: true) {
                         arState.actionStream.send(.fixBoardPosition)
                         arState.actionStream.send(.disablePlaneVisualization)
-                        arState.currentState = .boardConfirmed
+                        arState.gamePhase = .boardConfirmed
                     }
 
-                // 5. 보드가 확정된 상태 (게임 시작 대기)
+                // 4. 보드가 확정된 상태 (게임 시작 대기)
                 case .boardConfirmed:
                     EmptyView() // 상단 안내 없음
                     Spacer()
                     RoundedBrownButton(title: "윷놀이 시작!", isEnabled: true) {
-                        arState.currentState = .readyToThrow
+                        arState.gamePhase = .readyToThrow
                     }
 
-                // 6. 윷 던지기 준비 단계
+                // 5. 윷 던지기 준비 단계
                 case .readyToThrow:
                     InstructionView(text: "버튼을 눌러 윷을 던져랏")
                     Spacer()
@@ -73,7 +64,7 @@ struct ContentView : View {
                         arState.actionStream.send(.startMonitoringMotion)
                     }
 
-                // 7. 움직일 말 선택 단계
+                // 6. 움직일 말 선택 단계
                 case .selectingPieceToMove:
                     VStack {
                         InstructionView(text: "움직일 말을 선택하세요.")
@@ -99,7 +90,7 @@ struct ContentView : View {
                         }
                     }
 
-                // 8. 말의 이동 목적지 선택 단계
+                // 7. 말의 이동 목적지 선택 단계
                 case .selectingDestination:
                     InstructionView(text: "말을 옮길 곳을 선택하세요.")
                     Spacer()
