+++ Yut/Yut/Features/Lobby/Models/Room.swift
@@ -1,0 +1,6 @@
+struct Room: Identifiable, Hashable {
+    let id = UUID()
+    let name: String
+    let currentPlayers: Int
+    let maxPlayers: Int
+}
