+++ Yut/Yut/Features/Lobby/Views/NameInput/HostNameInputView.swift
@@ -1,10 +1,3 @@
-//
-//  HostNameInputView.swift
-//  Yut
-//
-//  Created by Hwnag Seyeon on 7/17/25.
-//
-
 import SwiftUI
 
 struct HostNameInputView: View {
